+++ atom/typed.py
@@ -5,7 +5,7 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 # --------------------------------------------------------------------------------------
-from .catom import DefaultValue, Member, Validate
+from .catom import DefaultValue, GetState, Member, Validate
 from .typing_utils import extract_types, is_optional
 
 
@@ -83,6 +83,8 @@ class Typed(Member):
             self.set_validate_mode(Validate.OptionalTyped, kind)
         else:
             self.set_validate_mode(Validate.Typed, kind)
+            # Allow to create a pickle with an unset typed value
+            self.set_getstate_mode(GetState.IncludeNonDefault, None)
 
 
 class ForwardTyped(Typed):
@@ -143,6 +145,9 @@ class ForwardTyped(Typed):
             if optional is not None
             else factory is None and args is None and kwargs is None
         )
+        if not self.optional:
+            # Allow to create a pickle with an unset typed value
+            self.set_getstate_mode(GetState.IncludeNonDefault, None)
 
         self.set_validate_mode(Validate.MemberMethod_ObjectOldNew, "validate")